refactor: MemberRepository
