+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.startingblue.fourtooncookie.member.domain;
+package com.startingblue.fourtooncookie.member;
 
+import com.startingblue.fourtooncookie.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
