+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.startingblue.fourtooncookie.member.service;
 
 import com.startingblue.fourtooncookie.member.domain.Member;
-import com.startingblue.fourtooncookie.member.domain.MemberRepository;
+import com.startingblue.fourtooncookie.member.MemberRepository;
 import com.startingblue.fourtooncookie.member.domain.Role;
 import com.startingblue.fourtooncookie.member.dto.request.MemberSaveRequest;
 import com.startingblue.fourtooncookie.member.exception.MemberDuplicateException;
