+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/diary/domain/DiaryTest.java
@@ -8,7 +8,7 @@ import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import com.startingblue.fourtooncookie.member.domain.Gender;
 import com.startingblue.fourtooncookie.member.domain.Member;
-import com.startingblue.fourtooncookie.member.domain.MemberRepository;
+import com.startingblue.fourtooncookie.member.MemberRepository;
 import com.startingblue.fourtooncookie.member.domain.Role;
 import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.Validation;
